Burp Extension: Simplify V0.5
This Burp extension is writen in python. It is triggered with the burp passive scan.  It simplifies the requests by checking what HTTP headers are actually needed to get the same request. This allows you to focus on the most relevant parts of the request. The output of this tools is also good for reporting a finding. Since it short and to the point.