Add files via upload
The param-test-v0.5.py Burp extension extracts from the HTTP Request the parameter names, header, file names and folder names. Then includes them in a text file that you can then use as a wordlist. It checks that the values are unique that it only contains the right characters and that it is no longer than 40 characters.